pyproject.toml
[project]
name = "superdoc-benchmark"
version = "0.1.0"
description = "Visual benchmarking tool for SuperDoc document rendering"
requires-python = ">=3.10"
dependencies = [
    "InquirerPy>=0.3.4",
    "rich>=13.0",
    "pymupdf>=1.24.0",
    "playwright>=1.40.0",
    "Pillow>=10.0.0",
]

[project.scripts]
superdoc-benchmark = "superdoc_benchmark.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/superdoc_benchmark"]

[tool.uv]
dev-dependencies = [
    "pyinstaller>=6.0",
]
